Minor VirtualDocument refactoring for better code reuse
