+++ src/Razor/src/Microsoft.VisualStudio.LanguageServer.ContainedLanguage/TextChangeExtensions.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using Microsoft.VisualStudio.LanguageServer.ContainedLanguage;
 using Microsoft.VisualStudio.Text;
-using RoslynTextChange = Microsoft.CodeAnalysis.Text.TextChange;
 
-namespace Microsoft.VisualStudio.LanguageServerClient.Razor
+namespace Microsoft.VisualStudio.LanguageServer.ContainedLanguage
 {
     public static class TextChangeExtensions
     {
@@ -25,8 +21,5 @@ namespace Microsoft.VisualStudio.LanguageServerClient.Razor
         {
             return textChange.OldSpan.Length > 0 && textChange.NewText.Length > 0;
         }
-
-        public static ITextChange ToVisualStudioTextChange(this RoslynTextChange roslynTextChange) =>
-            new VisualStudioTextChange(roslynTextChange.Span.Start, roslynTextChange.Span.Length, roslynTextChange.NewText);
     }
 }
