+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/HtmlVirtualDocument.cs
@@ -2,91 +2,17 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using Microsoft.VisualStudio.LanguageServer.ContainedLanguage;
 using Microsoft.VisualStudio.Text;
 
 namespace Microsoft.VisualStudio.LanguageServerClient.Razor
 {
-    internal class HtmlVirtualDocument : VirtualDocument
+    internal class HtmlVirtualDocument : VirtualDocumentBase<HtmlVirtualDocumentSnapshot>
     {
-        private long? _hostDocumentSyncVersion;
-        private HtmlVirtualDocumentSnapshot _currentSnapshot;
-
-        public HtmlVirtualDocument(Uri uri, ITextBuffer textBuffer)
+        public HtmlVirtualDocument(Uri uri, ITextBuffer textBuffer) : base(uri, textBuffer)
         {
-            if (uri is null)
-            {
-                throw new ArgumentNullException(nameof(uri));
-            }
-
-            if (textBuffer is null)
-            {
-                throw new ArgumentNullException(nameof(textBuffer));
-            }
-
-            Uri = uri;
-            TextBuffer = textBuffer;
-            _currentSnapshot = UpdateSnapshot();
         }
 
-        public override Uri Uri { get; }
-
-        public override ITextBuffer TextBuffer { get; }
-
-        public override long? HostDocumentSyncVersion => _hostDocumentSyncVersion;
-
-        public override VirtualDocumentSnapshot CurrentSnapshot => _currentSnapshot;
-
-        public override VirtualDocumentSnapshot Update(IReadOnlyList<ITextChange> changes, long hostDocumentVersion)
-        {
-            if (changes is null)
-            {
-                throw new ArgumentNullException(nameof(changes));
-            }
-
-            _hostDocumentSyncVersion = hostDocumentVersion;
-            TextBuffer.SetHostDocumentSyncVersion(_hostDocumentSyncVersion.Value);
-
-            if (changes.Count == 0)
-            {
-                // Even though nothing changed here, we want the synchronizer to be aware of the host document version change.
-                // So, let's make an empty edit to invoke the text buffer Changed events.
-                TextBuffer.MakeEmptyEdit();
-
-                _currentSnapshot = UpdateSnapshot();
-                return _currentSnapshot;
-            }
-
-            using var edit = TextBuffer.CreateEdit(EditOptions.None, reiteratedVersionNumber: null, InviolableEditTag.Instance);
-            for (var i = 0; i < changes.Count; i++)
-            {
-                var change = changes[i];
-
-                if (change.IsDelete())
-                {
-                    edit.Delete(change.OldSpan.Start, change.OldSpan.Length);
-                }
-                else if (change.IsReplace())
-                {
-                    edit.Replace(change.OldSpan.Start, change.OldSpan.Length, change.NewText);
-                }
-                else if (change.IsInsert())
-                {
-                    edit.Insert(change.OldSpan.Start, change.NewText);
-                }
-                else
-                {
-                    throw new InvalidOperationException("Unknown edit type when updating LSP HTML buffer.");
-                }
-            }
-
-            edit.Apply();
-            _currentSnapshot = UpdateSnapshot();
-
-            return _currentSnapshot;
-        }
-
-        private HtmlVirtualDocumentSnapshot UpdateSnapshot() => new HtmlVirtualDocumentSnapshot(Uri, TextBuffer.CurrentSnapshot, HostDocumentSyncVersion);
+        protected override HtmlVirtualDocumentSnapshot GetUpdatedSnapshot() => new HtmlVirtualDocumentSnapshot(Uri, TextBuffer.CurrentSnapshot, HostDocumentSyncVersion);
     }
 }
